--- a/demo/res/provider/app.json
+++ b/demo/res/provider/app.json
@@ -5,7 +5,7 @@
         "url": "https://cdn.openfin.co/services/openfin/test/provider.html",
         "uuid": "test-service",
         "autoShow": true,
-        "showTaskbarIcon": false,
+        "showTaskbarIcon": true,
         "defaultHeight": 400,
         "defaultWidth": 500
     },