--- a/demo/tsconfig.json
+++ b/demo/tsconfig.json
@@ -4,7 +4,10 @@
         "target": "es2017",
         "outDir": "dist",
         "rootDir": "src",
-        "lib": ["es2017"],
+        "lib": [
+            "es2017",
+            "dom"
+        ],
         "sourceMap": true,
         "declaration": false,
         "moduleResolution": "node"