--- a/package.json
+++ b/package.json
@@ -45,6 +45,8 @@
     "execa": "^1.0.0",
     "express": "^4.17.1",
     "file-loader": "^3.0.1",
+    "fork-ts-checker-notifier-webpack-plugin": "^1.0.2",
+    "fork-ts-checker-webpack-plugin": "^1.5.1",
     "jest": "^24.8.0",
     "jest-circus": "^24.8.0",
     "jest-cli": "^24.8.0",
@@ -59,13 +61,13 @@
     "sass-loader": "^7.1.0",
     "style-loader": "^1.0.0",
     "ts-jest": "^24",
-    "ts-loader": "^4.4.2",
+    "ts-loader": "^6.2.0",
     "typedoc": "^0.15.0",
     "typedoc-plugin-external-module-name": "^2.1.0",
     "url-loader": "^1.0.1",
-    "webpack": "^4.36.0",
+    "webpack": "^4.41.2",
     "webpack-cli": "^3.3.6",
-    "webpack-dev-middleware": "^3.7.0",
+    "webpack-dev-middleware": "^3.7.2",
     "ws": "^7.1.2"
   },
   "devDependencies": {
@@ -85,7 +87,7 @@
     "openfin-service-config": "^1.0.1",
     "openfin-service-signal": "^1.0.0",
     "pre-commit": "^1.2.2",
-    "typescript": "^3.5.3"
+    "typescript": "^3.6.4"
   },
   "peerDependencies": {
     "typescript": "3.x"