This PR optimizes service-tooling's build/watch performance, just from tests there's a 60% decrease in build time. This is achieved by splitting typescript compilation in to 2 steps: type-checking and transpiling. `ts-loader` will only transpile files without type checking now, and `fork-ts-checker` runs on a separate process and does the type-checking.

Try it out and see if there's an improvement on your end.

# Builds (FDC3)
<details>
 <summary>4 core Surface</summary>
Current
<pre>
Initial build complete after 64.408 seconds
  dist\client: 38.996s
  dist\client: 44.01s
  dist\provider: 14.206s
  dist\provider\ui: 7.576s
  dist\demo: 62.83s
</pre>
New
<pre>
Initial build complete after 27.595 seconds
   dist\client: 26.459s
   dist\client: 27.02s
   dist\provider: 27.613s
   dist\provider\ui: 26.781s
   dist\demo: 26.585s
</pre>
</details>

<details>
 <summary>8 core machine</summary>
Current
<pre>
Initial build complete after 22.707 seconds
  dist\client: 11.721s
  dist\client: 12.795s
  dist\provider: 3.393s
  dist\provider\ui: 4.483s
  dist\demo: 18.037s
</pre>
New
<pre>
Initial build complete after 8.034 seconds
  dist\client: 7.913s
  dist\client: 7.685s
  dist\provider: 7.999s
  dist\provider\ui: 7.881s
  dist\demo: 7.822s
</pre>
</details>