@bryangaleOF 2019-10-18
Outside of Parallels, build time goes from ~15s to ~10s for me